backend/pyproject.toml
[tool.poetry]
name = "spotify-recs-backend"
version = "0.1.0"
description = "Spotify playlist recommender backend (FastAPI) using raw Spotify Web API"
authors = ["You"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = {extras=["standard"], version="0.22.0"}
pandas = "^2.1.0"
scikit-learn = "^1.2.2"
requests = "^2.31.0"
python-dotenv = "^1.1.0"
pydantic = "^1.10.7"
redis = "^4.5.0"
rq = "^1.16.0"
psycopg2-binary = "^2.9.7"
SQLAlchemy = "^2.0.20"
pgvector = "^0.5.2"
sqlalchemy-utils = "^0.39.0"
tenacity = "^8.2.2"
